# Optimize ReadBlob and WriteBlob