--- a/server/backends/blobstore/disk/BUILD
+++ b/server/backends/blobstore/disk/BUILD
@@ -22,6 +22,7 @@ go_test(
     deps = [
         "//server/backends/blobstore/util",
         "//server/interfaces",
+        "//server/testutil/testdigest",
         "@com_github_stretchr_testify//require",
     ],
 )