--- a/server/backends/blobstore/disk/disk_test.go
+++ b/server/backends/blobstore/disk/disk_test.go
@@ -2,11 +2,14 @@ package disk
 
 import (
 	"context"
+	"fmt"
+	"math/rand"
 	"path/filepath"
 	"testing"
 
 	"github.com/buildbuddy-io/buildbuddy/server/backends/blobstore/util"
 	"github.com/buildbuddy-io/buildbuddy/server/interfaces"
+	"github.com/buildbuddy-io/buildbuddy/server/testutil/testdigest"
 	"github.com/stretchr/testify/require"
 )
 
@@ -98,6 +101,60 @@ func TestDiskBlobStore(t *testing.T) {
 			exist, err = bs.BlobExists(ctx, tc.blobName)
 			require.NoError(t, err)
 			require.False(t, exist)
+		})
+	}
+}
+
+type namedBlob struct {
+	name string
+	buf  []byte
+}
+
+func getBlobs(t testing.TB, num int, sizeBytes int64) []*namedBlob {
+	blobs := make([]*namedBlob, 0, num)
+	for i := 0; i < num; i++ {
+		d, b := testdigest.RandomCASResourceBuf(t, sizeBytes)
+		blobs = append(blobs, &namedBlob{
+			name: d.GetDigest().GetHash(),
+			buf:  b,
+		})
+	}
+	return blobs
+}
+
+func writeBlobs(t testing.TB, ctx context.Context, bs interfaces.Blobstore, blobs []*namedBlob) {
+	for _, b := range blobs {
+		if _, err := bs.WriteBlob(ctx, b.name, b.buf); err != nil {
+			t.Fatal(err)
+		}
+	}
+}
+
+func BenchmarkReadBlob(b *testing.B) {
+	originalRootDir := *rootDirectory
+	*rootDirectory = b.TempDir()
+	b.Cleanup(func() {
+		*rootDirectory = originalRootDir
+	})
+	var bs interfaces.Blobstore
+	bs, err := NewDiskBlobStore()
+	ctx := context.Background()
+	require.NoError(b, err)
+	for _, size := range []int64{10, 1e3, 1e4, 1e5, 1e6} {
+		blobs := getBlobs(b, 100, size)
+		writeBlobs(b, ctx, bs, blobs)
+		name := fmt.Sprintf("size=%d", size)
+		b.Run(name, func(b *testing.B) {
+			b.ReportAllocs()
+			b.ResetTimer()
+			for i := 0; i < b.N; i++ {
+				namedBlob := blobs[rand.Intn(len(blobs))]
+				b.SetBytes(int64(len(namedBlob.buf)))
+				_, err = bs.ReadBlob(ctx, namedBlob.name)
+				if err != nil {
+					b.Fatal(err)
+				}
+			}
 		})
 	}
 }