--- a/server/backends/blobstore/util/util.go
+++ b/server/backends/blobstore/util/util.go
@@ -49,12 +49,10 @@ func Decompress(in []byte, err error) ([]byte, error) {
 	if err != nil {
 		return nil, err
 	}
+	defer zr.Close()
 	var buffer bytes.Buffer
 	_, err = io.Copy(&buffer, zr)
 	if err != nil {
-		return nil, err
-	}
-	if err := zr.Close(); err != nil {
 		return nil, err
 	}
 	return buffer.Bytes(), nil
@@ -69,7 +67,8 @@ func Compress(in []byte) ([]byte, error) {
 	if err := zr.Close(); err != nil {
 		return nil, err
 	}
-	return io.ReadAll(&buf)
+
+	return buf.Bytes(), nil
 }
 
 // prefixBlobstore implements interfaces.Blobstore, is a wrapper around